+++ test/dummy/config/initializers/recording_studio_duplicatable.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+RecordingStudioDuplicatable.configure do |config|
+  config.duplication_prefix = nil
+  config.duplication_suffix = nil
+end