+++ test/dummy/config/routes.rb
@@ -7,6 +7,7 @@ Rails.application.routes.draw do
   mount RecordingStudioAccessible::Engine, at: "/admin/access"
   mount RecordingStudioAttachable::Engine, at: "/recording_studio_attachable"
   mount RecordingStudioPages::Engine, at: "/recording_studio_pages"
+  mount RecordingStudioDuplicatable::Engine, at: "/recording_studio_duplicatable"
   mount RecordingStudioPublishable::Engine, at: "/"
   recording_studio_admin_for :admin, at: "/admin", root_section: :pages